BOJ#2420. 사파리 월드

+++ BOJ/src/Main_2420.java
@@ -4,8 +4,8 @@ public class Main_2420 {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		long a = sc.nextLong();
+		long b = sc.nextLong();
 		System.out.println(Math.abs(a-b));
 		sc.close();
 	}
